requirements.txt
astroid==2.12.10
autopep8==1.7.0
bcc==0.18.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.3
defer==1.0.6
dill==0.3.5.1
distro==1.7.0
evdev==1.4.0
fonttools==4.38.0
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isort==5.10.1
jeepney==0.7.1
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
matplotlib==3.6.1
mccabe==0.7.0
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
numpy==1.23.4
oauthlib==3.2.0
olefile==0.46
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.2
pop-transition==1.1.2
protobuf==3.12.4
psycopg2-binary==2.9.5
pycairo==1.20.1
pycodestyle==2.9.1
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.15.3
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.1.1
repoman==1.4.0
requests==2.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
systemd-python==234
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
webencodings==0.5.1
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
